backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.3.2
